requirements.txt
drf-generators
djangorestframework
drf-yasg
pymysql
django-allauth
django-rest-auth
djangorestframework_jwt